Running_locally_code/requirements.txt
confluent-kafka
pyshark
PyYAML
